The push was rejected because another session had advanced the same branch by 16
commits, including 03a2ad44e "integrate Tensors 0.121.0 (#1946)" — a deliberate
integration of the same package bump I had hand-resolved locally against master.

Reassuringly, both arrived at the SAME version matrix (Tensors 0.121.0, OneDNN
0.120.6, OpenBLAS 0.121.0, CLBlast 0.120.6), so the max-of-each reasoning in the
previous merge held up against an independent integration.

The only conflict was comment prose. Both notes are kept because they document
different bumps and do not contradict: mine explains why OneDNN/CLBlast stay at
0.120.6 rather than following master back to 0.119.0 (#906 saved-state pinning,
#907 LSTM GPU kernels), theirs explains the 0.121.0 move (#925, full-span autodiff
views returned to the pool while the graph still owns them).

Verified: dotnet restore exits 0 with zero errors.

+++ src/SurvivalAnalysis/NelsonAalenEstimator.cs
@@ -129,7 +129,9 @@ public class NelsonAalenEstimator<T> : SurvivalModelBase<T>
                 cumulativeH += hazardIncrement;
 
                 // Variance increment: d(t)/n(t)^2
-                cumulativeVar += (double)eventsAtTime / (atRisk * atRisk);
+                // atRisk * atRisk is squared in int and overflows for a large cohort,
+                // so the variance term flips sign rather than shrinking.
+                cumulativeVar += (double)eventsAtTime / ((double)atRisk * atRisk);
             }
 
             _cumulativeHazard[t] = NumOps.FromDouble(cumulativeH);
